arm64: dts: sunxi: enable USB controllers and USB PHY nodes for Pine64
The Pine64 board has two USB Type-A ports, the lower one being connected
to the host only USB PHY port, the upper one to the OTG/HCI port.
Since a Type-A receptable lacks the OTG detection pin, we use the upper
port in permanent host mode, since this is what users would expect.

[Andre: Merge OTG patch and reword commit message]

Signed-off-by: Icenowy Zheng <icenowy@aosc.xyz>
Signed-off-by: Andre Przywara <andre.przywara@arm.com>