fix: pci: advk: correct PCIe max read request size setting
There is an obvious typo issue when setting PCIe max
read request size according to spec, but it does not
cause any functional issue, so it is only a typo issue.

The patch added the missed macro definetion of PCIe
max read request size, and correct its setting in
the driver.

Change-Id: I80bdd7b76a680b4f5863451d3a28b62fb66a4ba7
Signed-off-by: Evan Wang <xswang@marvell.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/39256
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Omri Itach <omrii@marvell.com>
Reviewed-by: Victor Gu <xigu@marvell.com>